+++ babel.config.js
@@ -2,8 +2,8 @@ module.exports = (api) => {
     api.cache(true);
 
     const presets = [
+        '@babel/preset-react',
         '@babel/preset-env',
-        '@babel/react',
     ];
     const plugins = [
         '@babel/plugin-syntax-dynamic-import',