+++ src/templates/drawer.jsx
@@ -66,6 +66,26 @@ const TRANSLATE_X_RIGHT_START = 'translateX(100%)';
 const BOX_SHADOW_XSMALL = '0 0 3px 0 rgba(0, 0, 0, 0.30)';
 const BOX_SHADOW_SMALL = '2px 0 7px 0 rgba(0, 0, 0, 0.17)';
 const BOX_SHADOW_LARGE = '12px 0 19px 0 rgba(0, 0, 0, .22)';
+const DEFAULT_DRAWER_WIDTH = 768;
+
+function toggleBodyStyle(isOpen) {
+    const { scrollY } = window;
+    const bodyStyleTop = BODY.style.top;
+
+    BODY.style.height = isOpen ? '100%' : '';
+    BODY.style.position = isOpen ? 'fixed' : '';
+    BODY.style.top = isOpen ? `-${scrollY}px` : '';
+    BODY.style.width = isOpen ? '100%' : '';
+
+    if (!isOpen) {
+        BODY.classList.remove('drawer-open');
+
+        // eslint-disable-next-line radix
+        window.scrollTo(0, parseInt(bodyStyleTop || '0') * -1);
+    } else {
+        BODY.classList.add('drawer-open');
+    }
+}
 
 class Drawer extends React.Component {
     constructor(props) {
@@ -118,23 +138,22 @@ class Drawer extends React.Component {
                 `${TRANSLATE_X_END} translateY(${nextProps.positionYOffset}px)` :
                 TRANSLATE_X_END;
         }
-        
+
         if (
             prevProps.isOpen &&
             nextProps.isOpen &&
             prevProps.maxWidth !== nextProps.maxWidth
         ) {
             this.drawerContainerRef.style.maxWidth = _.isNumber(nextProps.maxWidth) ? `${nextProps.maxWidth}px` :
-                nextProps.maxWidth || '768px';
+                nextProps.maxWidth || `${DEFAULT_DRAWER_WIDTH}px`;
         }
-        
+
         if (
             prevProps.isOpen &&
             nextProps.isOpen &&
             prevProps.maxHeight !== nextProps.maxHeight
         ) {
-            this.drawerContainerRef.style.maxHeight = _.isNumber(nextProps.maxHeight) ? `${nextProps.maxHeight}px` :
-                nextProps.maxHeight || '700px';
+            this.drawerContainerRef.style.maxHeight = nextProps.maxHeight ? `${nextProps.maxHeight}px` : 'none';
         }
     }
 
@@ -146,8 +165,10 @@ class Drawer extends React.Component {
          * closing animation never happens.
          */
         if (isOpen && this.useComponentWillUnmount) {
-            if (BODY.classList.contains('drawer-open')) {
-                BODY.classList.remove('drawer-open');
+            toggleBodyStyle(false);
+
+            if (BODY.classList.contains('drawer-nubbin-open')) {
+                BODY.classList.remove('drawer-nubbin-open');
             }
 
             if (BODY.classList.contains('drawer-dimmers')) {
@@ -212,14 +233,20 @@ class Drawer extends React.Component {
             BODY.classList.remove('drawer-open-layered');
         }
 
-        if (drawerLength <= 1) {
-            const scrollPosition = parseInt(BODY.style.top, 10);
-            window.scroll(0, Math.abs(scrollPosition));
-            BODY.classList.remove('drawer-open', 'drawer-dimmers');
-            BODY.style.top = null;
+        if (drawerLength <= 1 || (drawerLength > 1 && BODY.classList.contains('drawer-nubbin-open'))) {
+            const isOpen = false;
+
+            toggleBodyStyle(isOpen);
+
+            if (BODY.classList.contains('drawer-dimmers')) {
+                BODY.classList.remove('drawer-dimmers');
+            }
         }
 
-        BODY.style.position = null;
+        if (drawerLength <= 1 && BODY.classList.contains('drawer-nubbin-open')) {
+            BODY.classList.remove('drawer-nubbin-open');
+        }
+
         BODY.classList.remove('drawer-animate-out');
 
         this.drawerContainerRef.style.transform = this.setStartOfTransform();
@@ -252,7 +279,6 @@ class Drawer extends React.Component {
         const boxShadowPositionX = this.isPositionX('right') ? '-' : '';
         const drawerLength = document.querySelectorAll('.ui.drawer').length;
         const layeredOffset = 11;
-        const scrollPosition = window.pageYOffset;
         const zIndex = 10002; // adding 2 accounts for the frist .drawer and .drawer-dimmers- z-indexes
 
         this.drawerContainerRef.addEventListener(animationEvent, this.onOpenAnimationComplete);
@@ -312,12 +338,13 @@ class Drawer extends React.Component {
 
                     default:
                 }
-                BODY.style.top = `-${scrollPosition}px`;
 
-                domUtils.addClassName(BODY, 'drawer-open');
+                if (!positionY && !maxHeight) {
+                    const isOpen = true;
 
-                if (positionY && maxHeight) {
-                    BODY.style.position = 'inherit';
+                    toggleBodyStyle(isOpen);
+                } else {
+                    BODY.classList.add('drawer-nubbin-open');
                 }
 
                 this.shadowRef.style.boxShadow = `${boxShadowPositionX}${boxShadow}`;
@@ -327,14 +354,14 @@ class Drawer extends React.Component {
 
             if (!_.isUndefined(maxWidth)) {
                 this.drawerContainerRef.style.maxWidth = _.isNumber(maxWidth) ? `${maxWidth}px` :
-                    maxWidth || '768px';
+                    maxWidth || `${DEFAULT_DRAWER_WIDTH}px`;
             } else {
                 this.drawerContainerRef.style.maxWidth =
-                    `${768 - (layeredOffset * (drawerLength - 1))}px`;
+                    `${DEFAULT_DRAWER_WIDTH - (layeredOffset * (drawerLength - 1))}px`;
             }
 
             if (!_.isUndefined(maxHeight)) {
-                this.drawerContainerRef.style.maxHeight = maxHeight ? `${maxHeight}px` : '700px';
+                this.drawerContainerRef.style.maxHeight = `${maxHeight}px`;
             }
 
             this.drawerContainerRef.style.transform = _.isNumber(positionYOffset) ?
@@ -385,6 +412,7 @@ class Drawer extends React.Component {
             children,
             className,
             positionYOffset,
+            style,
             wing,
         } = this.props;
         const { isOpen } = this.state;
@@ -410,6 +438,7 @@ class Drawer extends React.Component {
                         ref={(ref) => { this.drawerContainerRef = ref; }}
                         style={{
                             height: _.isNumber(positionYOffset) ? `calc(100% - ${positionYOffset}px)` : null,
+                            ...style,
                         }}
                     >
                         <ScrollBar